moved ticket views and functions to ext
